mcp-server/pyproject.toml
[project]
name = "fin-regbase-mcp"
version = "0.1.0"
description = "MCP server exposing the fin-regbase financial regulations wiki as agent tools"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
    "rank-bm25>=0.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
